requirements.txt
asgiref==3.5.2
Django==4.1
django-discover-runner==1.0
django-role-permissions==3.1.1
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2